remove v2 in the file name
