+++ apps/admin_panel/assets/src/omg-page-each-account/AccountLayout.js
